New Update Method for Location and Location Type
git-svn-id: http://svn.encuesta.me/trunk/encuestame@311 39e2479b-dd58-4041-9f67-69c9b71f0610

+++ source/java/org/encuestame/core/persistence/dao/CatLocationDao.java
@@ -21,6 +21,7 @@ import java.util.List;
 
 import org.encuestame.core.persistence.dao.imp.ICatLocation;
 import org.encuestame.core.persistence.pojo.CatLocation;
+import org.encuestame.core.persistence.pojo.SecGroups;
 import org.hibernate.HibernateException;
 
 /**
@@ -70,9 +71,10 @@ public class CatLocationDao extends AbstractHibernateDaoSupport implements ICatL
     @SuppressWarnings("unchecked")
     public List<CatLocation> getLocationbyLevelId(final Long locateId) throws HibernateException{
         return getSession().createQuery("FROM CatLocation WHERE locationLevel = :locateId")
-        .setParameter("locateId", locateId)
+        .setParameter("locateId", Integer.valueOf(locateId.intValue()))
         .list();
 
 
     }
+
     }
